rms classes/emd, moving to classes/
