+++ py4DSTEM/io/classes/metadata.py
@@ -3,7 +3,7 @@ from numbers import Number
 from typing import Optional
 import h5py
 
-from py4DSTEM.io.classes.emd.tree import Tree
+from py4DSTEM.io.classes.tree import Tree
 
 
 class Metadata:
@@ -82,11 +82,11 @@ class Metadata:
     # HDF5 read/write
 
     def to_h5(self,group):
-        from py4DSTEM.io.classes.emd.io import Metadata_to_h5
+        from py4DSTEM.io.classes.io import Metadata_to_h5
         Metadata_to_h5(self,group)
 
     def from_h5(group):
-        from py4DSTEM.io.classes.emd.io import Metadata_from_h5
+        from py4DSTEM.io.classes.io import Metadata_from_h5
         return Metadata_from_h5(group)
