+++ py4DSTEM/io/classes/py4dstem/braggvectors.py
@@ -5,9 +5,9 @@ from typing import Optional,Union
 import numpy as np
 import h5py
 
-from py4DSTEM.io.classes.emd import PointListArray
-from py4DSTEM.io.classes.emd.tree import Tree
-from py4DSTEM.io.classes.emd.metadata import Metadata
+from py4DSTEM.io.classes import PointListArray
+from py4DSTEM.io.classes.tree import Tree
+from py4DSTEM.io.classes.metadata import Metadata
