+++ py4DSTEM/io/classes/py4dstem/braggvectors_fns.py
@@ -1,7 +1,7 @@
 # BraggVectors methods
 
 import numpy as np
-from py4DSTEM.io.classes.emd import Metadata
+from py4DSTEM.io.classes import Metadata
 
 
 # Bragg vector maps
