+++ py4DSTEM/io/classes/py4dstem/calibration.py
@@ -5,7 +5,7 @@ from numbers import Number
 from typing import Optional
 import h5py
 
-from py4DSTEM.io.classes.emd.metadata import Metadata
+from py4DSTEM.io.classes.metadata import Metadata
 from py4DSTEM.io.classes.py4dstem.propagating_calibration import propagating_calibration
 
 class Calibration(Metadata):
