+++ py4DSTEM/io/classes/py4dstem/probe_fns.py
@@ -1,7 +1,7 @@
 # Functions to become Probe methods
 
 import numpy as np
-from py4DSTEM.io.classes.emd import Metadata
+from py4DSTEM.io.classes import Metadata
