+++ py4DSTEM/io/classes/py4dstem/realslice.py
@@ -1,6 +1,6 @@
 # Defines the RealSlice class, which stores 2(+1)D real-space shaped data
 
-from py4DSTEM.io.classes.emd.array import Array
+from py4DSTEM.io.classes.array import Array
 
 from typing import Optional,Union
 import numpy as np
