+++ py4DSTEM/io/classes/root.py
@@ -3,7 +3,7 @@ from numbers import Number
 from typing import Optional
 import h5py
 
-from py4DSTEM.io.classes.emd.tree import Tree
+from py4DSTEM.io.classes.tree import Tree
 
 
 class Root:
@@ -51,11 +51,11 @@ class Root:
     # HDF5 read/write
 
     def to_h5(self,group):
-        from py4DSTEM.io.classes.emd.io import Root_to_h5
+        from py4DSTEM.io.classes.io import Root_to_h5
         Root_to_h5(self,group)
 
     def from_h5(group):
-        from py4DSTEM.io.classes.emd.io import Root_from_h5
+        from py4DSTEM.io.classes.io import Root_from_h5
         return Root_from_h5(group)
