+++ py4DSTEM/process/diskdetection/braggvectormap.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from py4DSTEM.process.utils import add_to_2D_array_from_floats
-from py4DSTEM.io.classes.emd import PointListArray
+from py4DSTEM.io.classes import PointListArray
 from py4DSTEM.utils.tqdmnd import tqdmnd
