+++ py4DSTEM/test/test_now/test_array.py
@@ -1,7 +1,7 @@
 # test import
 
 def test_Array():
-    from py4DSTEM.io.classes.emd import Array
+    from py4DSTEM.io.classes import Array
     import numpy as np
 
     shape = (3,4,5)
